+++ src/Exception/IncompleteUrlException.php
@@ -1,7 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Kuria\Url\Exception;
-
-class IncompleteUrlException extends \LogicException implements ExceptionInterface
-{
-}